#49 - Organizing screens in folders.
