+++ Routes.js
@@ -11,7 +11,7 @@ import UpdateInfoScreenContainer from './src/Containers/UpdateInfoScreenContaine
 import StockFoodCheckoutScreenContainer from './src/Containers/StockFoodCheckoutContainer';
 import SchedulingVisitContainer from './src/Containers/SchedulingVisitContainer';
 import SchoolInfoContainer from './src/Containers/SchoolInfoContainer';
-import MainReportsScreen from './src/screens/MainReportsScreen';
+import MainReportsScreen from './src/screens/reports/MainReportsScreen';
 import RefectoryCheckoutContainer from './src/Containers/RefectoryCheckoutContainer';
 import MenuContainer from './src/Containers/MenuContainer';
 import KitchenCheckoutContainer from './src/Containers/KitchenCheckoutContainer';
@@ -22,15 +22,15 @@ import SchoolSurroundingsCheckoutContainer from './src/Containers/SchoolSurround
 import FoodHandlerCheckoutContainer from './src/Containers/FoodHandlerCheckoutContainer';
 import WaterSewerSupplyCheckoutContainer from './src/Containers/WaterSewerSupplyCheckoutContainer';
 import FoodPreparationCheckoutContainer from './src/Containers/FoodPreparationCheckoutContainer';
-import PrincipalStartInspection from './src/screens/PrincipalStartInspection';
+import PrincipalStartInspection from './src/screens/startInspection/PrincipalStartInspection';
 import ScheduleMeetingContainer from './src/Containers/ScheduleMeetingContainer';
 import ScheduleMeetingMapContainer from './src/Containers/ScheduleMeetingMapContainer';
-import PrincipalManageRegister from './src/screens/PrincipalManageRegister';
-import MainLegislationScreen from './src/screens/MainLegislationScreen';
-import Resolution26Screen from './src/screens/Resolution26Screen';
-import Resolution1Screen from './src/screens/Resolution1Screen';
-import LawNumber11947Screen from './src/screens/LawNumber11947Screen';
-import LawNumber12982Screen from './src/screens/LawNumber12982Screen';
+import PrincipalManageRegister from './src/screens/manageCounselors/PrincipalManageRegister';
+import MainLegislationScreen from './src/screens/legislation/MainLegislationScreen';
+import Resolution26Screen from './src/screens/legislation/Resolution26Screen';
+import Resolution1Screen from './src/screens/legislation/Resolution1Screen';
+import LawNumber11947Screen from './src/screens/legislation/LawNumber11947Screen';
+import LawNumber12982Screen from './src/screens/legislation/LawNumber12982Screen';
 
 const Routes = () => ({
   render() {
@@ -117,10 +117,10 @@ const Routes = () => ({
             />
             <Scene key="schoolInfoScreen" component={SchoolInfoContainer} hideNavBar />
             <Scene key="lesgislationScreen" component={MainLegislationScreen} hideNavBar />
-            <Scene key="resolution26Screen" component={Resolution26Screen} hideNavBar />
-            <Scene key="resolution1Screen" component={Resolution1Screen} hideNavBar />
-            <Scene key="lawNumber11947Screen" component={LawNumber11947Screen} hideNavBar />
-            <Scene key="lawNumber12982Screen" component={LawNumber12982Screen} hideNavBar />
+            <Scene key="Resolution26Screen" component={Resolution26Screen} hideNavBar />
+            <Scene key="Resolution1Screen" component={Resolution1Screen} hideNavBar />
+            <Scene key="LawNumber11947Screen" component={LawNumber11947Screen} hideNavBar />
+            <Scene key="LawNumber12982Screen" component={LawNumber12982Screen} hideNavBar />
           </Scene>
         </Drawer>
       </Router>
