+++ __tests__/screens/MainReportsScreen.test.js
@@ -3,7 +3,7 @@ import Enzyme, { shallow } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import configureStore from 'redux-mock-store';
 // imported as a connected component!
-import MainReportsScreen from '../../src/screens/MainReportsScreen';
+import MainReportsScreen from '../../src/screens/reports/MainReportsScreen';
 
 Enzyme.configure({ adapter: new Adapter() });
