+++ __tests__/screens/PrincipalManageRegister.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Enzyme, { shallow } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 // import configureStore from 'redux-mock-store';
-import PrincipalManageRegister from '../../src/screens/PrincipalManageRegister';
+import PrincipalManageRegister from '../../src/screens/manageCounselors/PrincipalManageRegister';
 
 Enzyme.configure({ adapter: new Adapter() });
