+++ __tests__/screens/PrincipalStartInspection.test.js
@@ -1,9 +1,8 @@
 import React from 'react';
 import Enzyme, { shallow } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import configureStore from 'redux-mock-store';
 // import configureStore from 'redux-mock-store';
-import PrincipalStartInspection from '../../src/screens/PrincipalStartInspection';
+import PrincipalStartInspection from '../../src/screens/startInspection/PrincipalStartInspection';
 
 Enzyme.configure({ adapter: new Adapter() });
