+++ __tests__/screens/ScheduleMeeting.test.js
@@ -3,7 +3,7 @@ import Enzyme, { shallow } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import configureStore from 'redux-mock-store';
 import ScheduleMeetingContainer from '../../src/Containers/ScheduleMeetingContainer';
-import ScheduleMeeting from '../../src/screens/ScheduleMeeting';
+import ScheduleMeeting from '../../src/screens/scheduleMeeting/ScheduleMeeting';
 
 Enzyme.configure({ adapter: new Adapter() });
