+++ src/Containers/FoodQualityCheckoutContainer.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import FoodQualityCheckoutScreen from '../screens/FoodQualityCheckoutScreen';
+import FoodQualityCheckoutScreen from '../screens/reports/FoodQualityCheckoutScreen';
 import { setFoodQualityReportPositive, setFoodQualityReportNegative, setAcceptedMenu, setRefusedMenu, setFoodQualityObservation, setStatusFoodQuality } from '../actions/reportActions';
 
 const mapStateToProps = state => ({
