+++ src/Containers/KitchenCheckoutContainer.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import KitchenCheckoutScreen from '../screens/KitchenCheckoutScreen';
+import KitchenCheckoutScreen from '../screens/reports/KitchenCheckoutScreen';
 import { setKitchenReportPositive, setKitchenReportNegative, setKitchenObservation, setStatusKitchen } from '../actions/reportActions';
 
 const mapStateToProps = state => ({
