+++ src/Containers/ManageNotAcceptedRegistersScreenContainer.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import ManageNotAcceptedRegistersScreen from '../screens/ManageNotAcceptedRegistersScreen';
+import ManageNotAcceptedRegistersScreen from '../screens/manageCounselors/ManageNotAcceptedRegistersScreen';
 import { asyncGetCounselorFromGroup } from '../actions/listActions';
 import { disableCounselor, asyncAcceptCounselor } from '../actions/ManagerRegisterActions';
