+++ src/screens/legislation/Resolution26Screen.js
@@ -8,13 +8,13 @@ import {
 } from 'react-native';
 
 import Accordion from 'react-native-collapsible/Accordion';
-import RenderContent from '../components/AccordionContent';
-import RenderHeader from '../components/AccordionHeader';
-import Header from '../components/Header';
+import RenderContent from '../../components/AccordionContent';
+import RenderHeader from '../../components/AccordionHeader';
+import Header from '../../components/Header';
 import {
   CONTENT,
   SELECTORS,
-  RESOLUTION_1_DATE } from '../constants/legislation/resolutionNumber1';
+  RESOLUTION_26_DATE } from '../../constants/legislation/resolutionNumber26';
 
 const styles = StyleSheet.create({
   title: {
@@ -46,7 +46,7 @@ const styles = StyleSheet.create({
   },
 });
 
-export default class Resolution1Screen extends Component {
+export default class Resolution26Screen extends Component {
   state = {
     activeSection: false,
     collapsed: true,
@@ -63,10 +63,10 @@ export default class Resolution1Screen extends Component {
 
   render() {
     return (
-      <View style={{ flex: 1, backgroundColor: 'white' }}>
+      <View style={{ flex: 1, backgroundColor: 'white', paddingBottom: 20 }}>
         <Header
           title={'Legislação'}
-          subTitle={'Resolução Nº 1'}
+          subTitle={'Resolução Nº 26'}
           backButton
         />
         <ScrollView
@@ -91,7 +91,7 @@ export default class Resolution1Screen extends Component {
           </View>
         </ScrollView>
         <ScrollView>
-          <Text style={styles.title}>{RESOLUTION_1_DATE}</Text>
+          <Text style={styles.title}>{RESOLUTION_26_DATE}</Text>
           <Accordion
             activeSection={this.state.activeSection}
             sections={CONTENT}
