+++ src/screens/manageCounselors/ManageAcceptedRegistersScreen.js
