+++ src/screens/manageCounselors/ManageNotAcceptedRegistersScreen.js
