+++ src/screens/manageCounselors/PrincipalManageRegister.js
@@ -1,10 +1,9 @@
 import React, { Component } from 'react';
 import { StyleSheet } from 'react-native';
 import { TabViewAnimated, SceneMap, TabBar } from 'react-native-tab-view';
-import StartPendingInspectionContainer from '../Containers/StartPendingInspectionContainer';
-import StartExpiredInspectionContainer from '../Containers/StartExpiredInspectionContainer';
-import Header from '../components/Header';
-import StartAlreadyInspectionedInspectionContainer from '../Containers/StartAlreadyInspectionedInspectionContainer';
+import Header from '../../components/Header';
+import ManageNotAcceptedRegistersScreenContainer from '../../Containers/ManageNotAcceptedRegistersScreenContainer';
+import ManageAcceptedRegistersScreenContainer from '../../Containers/ManageAcceptedRegistersScreenContainer';
 
 const styles = StyleSheet.create({
   container: {
@@ -19,9 +18,8 @@ export default class Principal extends Component {
     this.state = {
       index: 0,
       routes: [
-        { key: 'first', title: 'Pendentes' },
-        { key: 'second', title: 'Fiscalizados' },
-        { key: 'third', title: 'Expirados' },
+        { key: 'first', title: 'Não gerenciados' },
+        { key: 'second', title: 'Validados' },
       ],
     };
   }
@@ -30,12 +28,11 @@ export default class Principal extends Component {
 
   renderFooter = props => <TabBar style={{ backgroundColor: '#FF9500', borderTopWidth: 1, borderBottomColor: 'black' }} tabStyle={{ paddingHorizontal: -0 }} {...props} />;
 
-  renderHeader = props => <Header title={'Visitas Agendadas'} backButton {...props} />
+  renderHeader = props => <Header title={'Gerenciar Registro'} backButton {...props} />
 
   renderScene = SceneMap({
-    first: StartPendingInspectionContainer,
-    second: StartAlreadyInspectionedInspectionContainer,
-    third: StartExpiredInspectionContainer,
+    first: ManageNotAcceptedRegistersScreenContainer,
+    second: ManageAcceptedRegistersScreenContainer,
   });
 
   render() {
